+++ flutter_inappwebview_android/lib/src/in_app_webview/in_app_webview.dart
@@ -1,16 +1,15 @@
 import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter_inappwebview_platform_interface/flutter_inappwebview_platform_interface.dart';
-import 'headless_in_app_webview.dart';
 
 import '../find_interaction/find_interaction_controller.dart';
-import 'in_app_webview_controller.dart';
-import '../pull_to_refresh/main.dart';
 import '../pull_to_refresh/pull_to_refresh_controller.dart';
+import 'headless_in_app_webview.dart';
+import 'in_app_webview_controller.dart';
 
 /// Object specifying creation parameters for creating a [PlatformInAppWebViewWidget].
 ///
@@ -128,6 +127,7 @@ class AndroidInAppWebViewWidgetCreationParams
       super.onCameraCaptureStateChanged,
       super.onMicrophoneCaptureStateChanged,
       super.onContentSizeChanged,
+      super.onShowFileChooser,
       super.initialUrlRequest,
       super.initialFile,
       super.initialData,
@@ -246,6 +246,7 @@ class AndroidInAppWebViewWidgetCreationParams
             onMicrophoneCaptureStateChanged:
                 params.onMicrophoneCaptureStateChanged,
             onContentSizeChanged: params.onContentSizeChanged,
+            onShowFileChooser: params.onShowFileChooser,
             initialUrlRequest: params.initialUrlRequest,
             initialFile: params.initialFile,
             initialData: params.initialData,
@@ -459,6 +460,10 @@ class AndroidInAppWebViewWidget extends PlatformInAppWebViewWidget {
     if (params.onNavigationResponse != null &&
         settings.useOnNavigationResponse == null) {
       settings.useOnNavigationResponse = true;
+    }
+    if (params.onShowFileChooser != null &&
+        settings.useOnShowFileChooser == null) {
+      settings.useOnShowFileChooser = true;
     }
   }