+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_headless_in_app_webview.dart
@@ -124,6 +124,7 @@ class PlatformHeadlessInAppWebViewCreationParams
       super.onContentSizeChanged,
       super.onProcessFailed,
       super.onAcceleratorKeyPressed,
+      super.onShowFileChooser,
       super.initialUrlRequest,
       super.initialFile,
       super.initialData,